ST-13, ST-10: fix review issues
