+++ backend/app/app/settings.py
@@ -27,8 +27,6 @@ DEBUG = True
 
 ALLOWED_HOSTS = []
 
-# AUTH_USER_MODEL = 'polls.User'
-LOGIN_REDIRECT_URL = '/'
 SESSION_COOKIE_HTTPONLY = False
 
 # Application definition
@@ -40,7 +38,7 @@ INSTALLED_APPS = [
     "django.contrib.sessions",
     "django.contrib.messages",
     "django.contrib.staticfiles",
-    "polls"
+    "core"
 ]
 
 MIDDLEWARE = [
