+++ backend/app/core/admin.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin
 
-from polls.models import User
+from core.models import User
 
 # Register your models here.
 admin.site.register(User, UserAdmin)
