+++ backend/app/core/service_layer.py
@@ -2,26 +2,20 @@ import pydantic
 from django.contrib.auth import authenticate, login
 from django.core.handlers.wsgi import WSGIRequest
 
-from polls.models import User
+from core.models import User
+from core.validators import LoginUserVO, LoginUserRequest, RegisterUserVO
 
 
-class RegisterUserVO(pydantic.BaseModel):
-    username: str
-    password: str
-
-
-class LoginUserVO(pydantic.BaseModel):
-    username: str
-    password: str
-
-
-def register_user_service(register_user_vo: RegisterUserVO):
+def register_user_service(request, register_user_vo: RegisterUserVO):
     user = User.objects.create_user(
         username=register_user_vo.username,
         password=register_user_vo.password
     )
     user.save()
-    return user
+    login_user_request = LoginUserRequest.model_validate_json(request.body)
+    login_vo = LoginUserVO(**login_user_request.model_dump())
+    login_user_response = login_user_service(request, login_vo)
+    return login_user_response
 
 
 def login_user_service(request: WSGIRequest, login_user_vo: LoginUserVO):
@@ -33,4 +27,4 @@ def login_user_service(request: WSGIRequest, login_user_vo: LoginUserVO):
     if user is not None:
         login(request, user)
         return {'user_id': user.id}
-    return None
+    return User.DoesNotExist
